Finished A3 Seq and started Cilk version
Was able to get the sequential version of the puzzle to work using bit
manipulation. Also started to work on the Cilk version but due to the
fact that Visual Studio and the Intel C++ compiler did not understand
cilk libraries/operators I was unable to work on that part.